feat: 상위 및 하위 지역코드 조회 API 구현

+++ src/main/java/com/oseak/myFestaBackend/repository/RegionRepository.java
@@ -6,6 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.oseak.myFestaBackend.entity.Region;
 
 @Repository
-public interface RegionRepository extends JpaRepository<Region, Long> {
-
+public interface RegionRepository extends JpaRepository<Region, Integer> {
 }
